fix breaking changes from recent serenity updates

+++ scripty_utils/src/latency/ws.rs
@@ -3,7 +3,7 @@ use serenity::client::bridge::gateway::ShardId;
 
 pub async fn get_ws_latency(
     mgr: &<ShardManagerWrapper as TypeMapKey>::Value,
-    shard_id: u64,
+    shard_id: u32,
 ) -> Option<u128> {
     let mgr_lock = mgr.lock().await;
     let runners = mgr_lock.runners.lock().await;
